+++ frontend/src/components/Posts/AllPosts.css
@@ -7,4 +7,5 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  font-family: sans-serif;
 }
