+++ frontend/src/components/SignupFormPage/SignUpForm.js
@@ -1,0 +1,101 @@
+import React, { useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { Redirect } from "react-router-dom";
+import * as sessionActions from "../../store/session";
+import './SignupForm.css';
+
+function SignupFormPage() {
+  const dispatch = useDispatch();
+  const sessionUser = useSelector((state) => state.session.user);
+  const [email, setEmail] = useState("");
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
+  const [errors, setErrors] = useState([]);
+
+  if (sessionUser) return <Redirect to="/" />;
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (password === confirmPassword) {
+      setErrors([]);
+      return dispatch(sessionActions.signup({ email, username, password }))
+        .catch(async (res) => {
+          const data = await res.json();
+          if (data && data.errors) setErrors(data.errors);
+        });
+    }
+    return setErrors(['Confirm Password field must be the same as the Password field']);
+  };
+
+  return (
+    <div className="signUpContainer">
+      <form onSubmit={handleSubmit}>
+        <ul>
+          {errors.map((error, idx) => <li key={idx}>{error}</li>)}
+        </ul>
+        <div>
+          <label>
+            Email
+          </label>
+        </div>
+        <div>
+          <input
+            type="text"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            required
+            className="inputField"
+          />
+        </div>
+        <div>
+          <label>
+            Username
+          </label>
+        </div>
+        <div>
+          <input
+            type="text"
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+            required
+            className="inputField"
+          />
+        </div>
+        <div>
+          <label>
+            Password
+          </label>
+        </div>
+        <div>
+          <input
+            type="password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            required
+            className="inputField"
+          />
+        </div>
+        <div>
+          <label>
+            Confirm Password
+          </label>
+        </div>
+        <div>
+          <input
+            type="password"
+            value={confirmPassword}
+            onChange={(e) => setConfirmPassword(e.target.value)}
+            required
+            className="inputField"
+          />
+        </div>
+        <div>
+          <button type="submit" className="signUpButton">Sign Up</button>
+        </div>
+      </form>
+    </div>
+  );
+}
+
+export default SignupFormPage;
