+++ frontend/src/components/SignupFormPage/SignupForm.css
@@ -1,0 +1,49 @@
+.signUpContainer {
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  /* margin: 20px; */
+}
+
+.inputField {
+  margin: 10px;
+}
+
+.inputField:hover {
+  border: blue;
+  border-style: solid;
+  border-radius: 1px;
+}
+
+.signUpButton {
+  background-color: #2e69ff;
+  transition-duration: 0.4s;
+  height: 30px;
+  min-width: 30px;
+  position: relative;
+  border-width: 0px;
+  color: white;
+  border-radius: 10px;
+  padding: 0px 20px;
+}
+
+.signUpButtonOnLogin {
+  background-color: transparent;
+  background-repeat: no-repeat;
+  border: none;
+  cursor: pointer;
+  overflow: hidden;
+  outline: none;
+}
+
+.signUpButtonOnLogin:hover {
+  background-color: lightgray;
+  border-radius: 10px;
+  transition-duration: 0.4s;
+  height: 30px;
+  min-width: 30px;
+  position: relative;
+  border-width: 0px;
+  /* padding: 0px 10px; */
+}
